+++ Assets/_Scripts/Utility/OverworldSceneChanger1.cs
@@ -19,13 +19,13 @@ public class OverworldSceneChanger1 : MonoBehaviour {
     private vThirdPersonController thirdPersonController;
     Animator animator;  
     
-    public int gabiCurrentHealth;
-    public int gabiCurrentResolve;
+    public float gabiCurrentHealth;
+    public float gabiCurrentResolve;
 
-    public int arvandusCurrentHealth;
-    public int arvanusCurrentStamina;
+    public float arvandusCurrentHealth;
+    public float arvanusCurrentStamina;
 
-    public int quinnCurrentHealth;
+    public float quinnCurrentHealth;
     public int quinnCurrentFire;
     public int quinnCurrentEarth;
     public int quinnCurrentWater;
